+++ parasol-knowledge/parasol-knowledge-provider/src/main/java/com/ginkgocap/parasol/knowledge/service/impl/KnowledgeService.java
@@ -57,7 +57,7 @@ public class KnowledgeService implements IKnowledgeService {
 		
 		long knowledgeId = this.knowledgeCommonService.getKnowledgeSeqenceId();
 		
-		long columnId = knowledgeMongo.getColumnId();
+		short columnId = knowledgeMongo.getColumnId();
 		
 		knowledgeMongo.setId(knowledgeId);
 		
@@ -350,7 +350,7 @@ public class KnowledgeService implements IKnowledgeService {
 	 * @date 2016年1月15日 上午11:30:18
 	 * @throws Exception
 	 */
-	private void insertRollBack(long knowledgeId, long columnId, Long userId,boolean isMongo,boolean isBase,boolean isReference,boolean isBigData,boolean isUserFeed) throws Exception {
+	private void insertRollBack(long knowledgeId, short columnId, Long userId,boolean isMongo,boolean isBase,boolean isReference,boolean isBigData,boolean isUserFeed) throws Exception {
 		if(isMongo) this.knowledgeMongoDao.deleteByIdAndColumnId(knowledgeId, columnId);
 		if(isBase) this.knowledgeBaseDao.deleteById(knowledgeId);
 		if(isReference) this.knowledgeReferenceDao.deleteByKnowledgeId(knowledgeId);
